Start simple `UserController` tests
This is a start on some basic JUnit tests for `UserController`. This took a little longer than I expected,
in part because I had to figure out how to parse JSON and deal with the resulting BSON documents,
and in part because I had to (OK, maybe chose to) learn how the new-ish (but new to me) Java 8
streams and collections worked.