Get `gradle` to deploy a stand-alone collection of artifacts
This does three things that collectively allow `gradle` to generate
a tarball or zipfile that can be moved to whatever machine/directory
you want to run the system on, and it will work when you untar/unzip
it (assuming Mongo is running on that machine). The first two are
specified in `server/build.gradle` and tell `gradle` where to put
the files and how to construct the associated CLASSPATH. The third
is in `Server.java` and tells Spark where to find those static files, i.e.,
how to map something like `host:4567/index.html` to the `index.html`
file generated by Angular.

* Put the client artifacts (HTML, CSS, and JS) generated by Angular in
   in `lib/client/public`.
* Add `lib/client` to the classpath put together by the startup script
   that `gradle` builds for us.
* Set `/public` as the directory that static files will be served from.
   Since that has to be on the classpath somewhere, putting `/public`
   in `lib/client` and adding `lib/client` to the classpath allows Spark
   to find the static files there.

The details of the directory structure are somewhat arbitrary so we can
change some of the names and locations if we want (and I don't make
any claims about the wonderfullness of my choices). As far as I understand
(and I don't claim to be terribly clear on any of this), it looks like anything
we put in the classpath needs to be in the `lib` directory. This is mildly
annoying; it would be nice to have the client artifacts _next_ to the
server artifacts instead of inside of them, but there you are. Beyond that,
though, my uses of `client` and `public` are pretty arbitrary.